// types/config.ts
export interface ConfiguracionHorario {
  director: string
  secretario: string
  recursosHumanos: string
  periodo: string
  fechaRevision: string
  numeroRevision: string
  codigoDocumento: string
  nombreUniversidad: string
  nombreDireccion: string
  direccionUniversidad: string
  logoUrl: string
  fecha?: string
  horasFrenteGrupo?: number
}